IngesoftBackend/requirements.txt
apicore==1.2.1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
et_xmlfile==2.0.0
Faker==30.8.1
filelock==3.16.1
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
platformdirs==4.3.6
pymongo==4.10.1
pystache==0.6.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.0.4
